third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-out-event-listener.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    function inner()
    {
        var div = document.getElementById("myDiv");
        function fooEventHandler1()
        {
            div.textContent += "Recieved foo event(1)!\n";
        }

        function fooEventHandler2()
        {
            div.textContent += "Recieved foo event(2)!\n";
        }

        div.addEventListener("foo", fooEventHandler1);
        div.addEventListener("foo", fooEventHandler2);

        var e = new CustomEvent("foo");
        debugger;
        div.dispatchEvent(e);

        div.removeEventListener("foo", fooEventHandler1);
        div.removeEventListener("foo", fooEventHandler2);
    }
    inner();
}

var test = function() {
  SourcesTestRunner.startDebuggerTest(step1, true);

  function step1() {
    SourcesTestRunner.runTestFunctionAndWaitUntilPaused(step2);
  }

  function step2() {
    var actions = [
      'Print',  // debugger;
      'StepInto',
      'StepInto',
      'StepInto',
      'Print',  // at fooEventHandler1
      'StepOut',
      'Print',  // should be at fooEventHandler2
      'StepOut',
      'Print',
    ];
    SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions, step3);
  }

  function step3() {
    SourcesTestRunner.resumeExecution(step4);
  }

  function step4() {
    SourcesTestRunner.runTestFunctionAndWaitUntilPaused(step5);
  }

  function step5() {
    var actions = [
      'Print',             // debugger;
      'StepOut', 'Print',  // should be at inner()
    ];
    SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions, step6);
  }

  function step6() {
    SourcesTestRunner.completeDebuggerTest();
  }
};

</script>
</head>

<body onload="runTest()">
<p>
Tests that stepping out of an event listener will lead to a pause in the next event listener.
</p>
<div id="myDiv"></div>
</body>
</html>
